+++ packages/manager/core/shell-client/src/hooks/useNavigation.ts
@@ -2,7 +2,6 @@ import { useShell } from './useShell';
 
 export const useNavigation = () => {
   const shell = useShell();
-
   return shell.navigation;
 };