+++ packages/manager/modules/nutanix/src/service.js
@@ -163,6 +163,12 @@ export default class NutanixService {
       .then(({ data }) => data);
   }
 
+  getNodeDetails(nodes) {
+    return this.$q
+      .all(nodes.map((node) => this.getServer(node.server)))
+      .then((res) => res);
+  }
+
   getServer(nodeId) {
     return this.$http
       .get(`/sws/dedicated/server/${nodeId}`, {