+++ packages/manager/apps/email-pro/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-emailpro-app@0.9.3...@ovh-ux/manager-emailpro-app@0.9.4) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-emailpro-app
+
+
+
+
+
 ## [0.9.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-emailpro-app@0.9.2...@ovh-ux/manager-emailpro-app@0.9.3) (2024-03-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-emailpro-app