+++ packages/manager/apps/octavia-load-balancer/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-octavia-load-balancer-app@0.3.8...@ovh-ux/manager-octavia-load-balancer-app@0.3.9) (2024-03-18)
+
+**Note:** Version bump only for package @ovh-ux/manager-octavia-load-balancer-app
+
+
+
+
+
+## [0.3.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-octavia-load-balancer-app@0.3.7...@ovh-ux/manager-octavia-load-balancer-app@0.3.8) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/manager-octavia-load-balancer-app
+
+
+
+
+
+## [0.3.7](https://github.com/ovh/manager/compare/@ovh-ux/manager-octavia-load-balancer-app@0.3.6...@ovh-ux/manager-octavia-load-balancer-app@0.3.7) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-octavia-load-balancer-app
+
+
+
+
+
+## [0.3.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-octavia-load-balancer-app@0.3.5...@ovh-ux/manager-octavia-load-balancer-app@0.3.6) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-octavia-load-balancer-app
+
+
+
+
+
+## [0.3.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-octavia-load-balancer-app@0.3.4...@ovh-ux/manager-octavia-load-balancer-app@0.3.5) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-octavia-load-balancer-app
+
+
+
+
+
 ## [0.3.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-octavia-load-balancer-app@0.3.3...@ovh-ux/manager-octavia-load-balancer-app@0.3.4) (2024-02-28)
 
 **Note:** Version bump only for package @ovh-ux/manager-octavia-load-balancer-app