+++ packages/manager/apps/order-tracking/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.6](https://github.com/ovh/manager/compare/@ovh-ux/order-tracking-app@2.8.5...@ovh-ux/order-tracking-app@2.8.6) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/order-tracking-app
+
+
+
+
+
+## [2.8.5](https://github.com/ovh/manager/compare/@ovh-ux/order-tracking-app@2.8.4...@ovh-ux/order-tracking-app@2.8.5) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/order-tracking-app
+
+
+
+
+
+## [2.8.4](https://github.com/ovh/manager/compare/@ovh-ux/order-tracking-app@2.8.3...@ovh-ux/order-tracking-app@2.8.4) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/order-tracking-app
+
+
+
+
+
+## [2.8.3](https://github.com/ovh/manager/compare/@ovh-ux/order-tracking-app@2.8.2...@ovh-ux/order-tracking-app@2.8.3) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/order-tracking-app
+
+
+
+
+
 ## [2.8.2](https://github.com/ovh/manager/compare/@ovh-ux/order-tracking-app@2.8.1...@ovh-ux/order-tracking-app@2.8.2) (2024-02-26)
 
 **Note:** Version bump only for package @ovh-ux/order-tracking-app