+++ packages/manager/apps/pci-rancher/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-rancher-app@1.1.3...@ovh-ux/manager-pci-rancher-app@1.1.4) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-rancher-app
+
+
+
+
+
+## [1.1.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-rancher-app@1.1.2...@ovh-ux/manager-pci-rancher-app@1.1.3) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-rancher-app
+
+
+
+
+
+## [1.1.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-rancher-app@1.1.1...@ovh-ux/manager-pci-rancher-app@1.1.2) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-rancher-app
+
+
+
+
+
+## [1.1.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-rancher-app@1.1.0...@ovh-ux/manager-pci-rancher-app@1.1.1) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-rancher-app
+
+
+
+
+
 # [1.1.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-rancher-app@1.0.0...@ovh-ux/manager-pci-rancher-app@1.1.0) (2024-02-28)