+++ packages/manager/apps/pci/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.11](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.1.10...@ovh-ux/manager-pci-app@4.1.11) (2024-03-18)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.1.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.1.9...@ovh-ux/manager-pci-app@4.1.10) (2024-03-13)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.1.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.1.8...@ovh-ux/manager-pci-app@4.1.9) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.1.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.1.7...@ovh-ux/manager-pci-app@4.1.8) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.1.7](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.1.6...@ovh-ux/manager-pci-app@4.1.7) (2024-03-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
 ## [4.1.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.1.5...@ovh-ux/manager-pci-app@4.1.6) (2024-03-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-app