+++ packages/manager/modules/hub/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.9.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub@3.9.0...@ovh-ux/manager-hub@3.9.1) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub
+
+
+
+
+
 # [3.9.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub@3.8.0...@ovh-ux/manager-hub@3.9.0) (2024-02-21)