+++ packages/manager/modules/models/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.11](https://github.com/ovh/manager/compare/@ovh-ux/manager-models@1.14.10...@ovh-ux/manager-models@1.14.11) (2024-03-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-models
+
+
+
+
+
 ## [1.14.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-models@1.14.9...@ovh-ux/manager-models@1.14.10) (2023-12-05)
 
 **Note:** Version bump only for package @ovh-ux/manager-models